nodes do reset when their frisbee session is over - not waiting for all nodes replace -1 -2 -3 with -n to skip reset (and no more stage3) frisbeed should be killed properly even with a timeout

+++ imaging/frisbeed.py
@@ -61,8 +61,12 @@ class Frisbeed:
 
     @asyncio.coroutine
     def stop(self):
-        self.subprocess.kill()
-        self.subprocess = None
-        logger.info("frisbeed stopped")
-        self.message_bus.put({'loader': "frisbee server stopped"})
-        pass
+        self.stop_wait()
+        
+    def stop_wait(self):
+        # make it idempotent
+        if self.subprocess:
+            self.subprocess.kill()
+            self.subprocess = None
+            logger.info("frisbeed stopped")
+            self.message_bus.put({'loader': "frisbee server stopped"})
